Extracting `SearchesDriver` from `Searches` class.
